% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOLB7950-package.R
\docType{package}
\name{MOLB7950}
\alias{MOLB7950}
\alias{MOLB7950-package}
\title{Informatics and Statistics for Molecular Biologists (MOLB 7950)}
\description{

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jayhesselberth/MOLB7900}
}

}
\author{
\strong{Maintainer}: Jay Hesselberth \email{jay.hesselberth@gmail.com}

Authors:
\itemize{
  \item Neel Mukherjee \email{neelanjan.mukherjee@ucdenver.edu}
}

}
